# Send preprocessed blobs to the master.