--- a/bdb/bdb_api.h
+++ b/bdb/bdb_api.h
@@ -706,7 +706,8 @@ int bdb_prim_allocdta_genid(bdb_state_type *bdb_handle, tran_type *tran,
                             int updateid, int *bdberr);
 int bdb_prim_adddta_n_genid(bdb_state_type *bdb_state, tran_type *tran,
                             int dtanum, void *dtaptr, size_t dtalen, int rrn,
-                            unsigned long long genid, int *bdberr);
+                            unsigned long long genid, int *bdberr,
+                            int odhready);
 
 int bdb_prim_deallocdta_genid(bdb_state_type *bdb_handle, tran_type *tran,
                               int rrn, unsigned long long genid, int *bdberr);
@@ -725,7 +726,7 @@ int bdb_prim_add_upd_genid(bdb_state_type *bdb_state, tran_type *tran,
                            int dtanum, void *newdta, int newdtaln, int rrn,
                            unsigned long long oldgenid,
                            unsigned long long newgenid, int participantstripeid,
-                           int *bdberr);
+                           int *bdberr, int odhready);
 
 int bdb_prim_no_upd(bdb_state_type *bdb_state, tran_type *tran, int rrn,
                     unsigned long long oldgenid, unsigned long long newgenid,
@@ -2124,4 +2125,11 @@ int truncate_asof_pglogs(bdb_state_type *bdb_state, int file, int offset);
 int bdb_set_logical_live_sc(bdb_state_type *bdb_state, int lock);
 int bdb_clear_logical_live_sc(bdb_state_type *bdb_state, int lock);
 
+/* Pack the payload into heap memory */
+int bdb_pack_heap(bdb_state_type *bdb_state, void *in, size_t inlen, void **out,
+                  size_t *outlen, void **freeptr);
+/* If the payload is uncompressed, point `out' to where the record starts;
+ * Otherwise unpack the payload into heap memory. */
+int bdb_unpack_heap(bdb_state_type *bdb_state, void *in, size_t inlen,
+                    void **out, size_t *outlen, void **freeptr);
 #endif