--- a/bdb/bdb_int.h
+++ b/bdb/bdb_int.h
@@ -1261,6 +1261,11 @@ int bdb_pack(bdb_state_type *bdb_state, const struct odh *odh, void *to,
 int bdb_unpack(bdb_state_type *bdb_state, const void *from, size_t fromlen,
                void *to, size_t tolen, struct odh *odh, void **freeptr);
 
+/* This is used by */
+int bdb_unpack_force_odh(bdb_state_type *bdb_state, const void *from,
+                         size_t fromlen, void *to, size_t tolen,
+                         struct odh *odh, void **freeptr);
+
 int bdb_retrieve_updateid(bdb_state_type *bdb_state, const void *from,
                           size_t fromlen);