--- a/bdb/odh.c
+++ b/bdb/odh.c
@@ -1091,9 +1091,9 @@ int bdb_get_unpack_blob(bdb_state_type *bdb_state, DB *db, DB_TXN *tid,
     return bdb_get_unpack_int(bdb_state, db, tid, key, data, ver, flags, 0);
 }
 
-static int bdb_prepare_put_pack_updateid(bdb_state_type *bdb_state, int is_blob,
-                                         DBT *data, DBT *data2, int updateid,
-                                         void **freeptr, void *stackbuf)
+int bdb_prepare_put_pack_updateid(bdb_state_type *bdb_state, int is_blob,
+                                  DBT *data, DBT *data2, int updateid,
+                                  void **freeptr, void *stackbuf, int odhready)
 {
     struct odh odh;
 
@@ -1101,29 +1101,24 @@ static int bdb_prepare_put_pack_updateid(bdb_state_type *bdb_state, int is_blob,
 
     memcpy(data2, data, sizeof(DBT));
 
-    init_odh(bdb_state, &odh, data->data, data->size, is_blob);
+    if (odhready) {
+        /* If the record is preprocessed on a replicant, use it as is. */
+        assert(is_blob);
+        *freeptr = NULL;
+        rc = 0;
+    } else {
+        init_odh(bdb_state, &odh, data->data, data->size, is_blob);
 
-    if (updateid > 0) {
-        odh.updateid = updateid;
+        if (updateid > 0) {
+            odh.updateid = updateid;
+        }
+
+        rc = bdb_pack(bdb_state, &odh, stackbuf, odh.length + ODH_SIZE_RESERVE,
+                      &data2->data, &data2->size, freeptr);
     }
 
-    /*
-    char s[128];
-    printf("Prepare to pack %u bytes with ODH [%s] is_blob=%d compress=%d:\n",
-    data->size,
-          snodhf(s, sizeof(s), &odh), is_blob, bdb_state->compress);
-    fsnapf(stdout, data->data, data->size);
-    */
-    rc = bdb_pack(bdb_state, &odh, stackbuf, odh.length + ODH_SIZE_RESERVE,
-                  &data2->data, &data2->size, freeptr);
-
-    if (rc == 0) {
-        /*
-        printf("Packed into %u bytes:\n", data2->size);
-        fsnapf(stdout, data2->data, data2->size);
-        */
+    if (rc == 0)
         data2->ulen = data2->size;
-    }
     return rc;
 }
 
@@ -1175,18 +1170,12 @@ int bdb_put(bdb_state_type *bdb_state, DB *db, DB_TXN *tid, DBT *key, DBT *data,
 }
 
 int bdb_put_pack(bdb_state_type *bdb_state, int is_blob, DB *db, DB_TXN *tid,
-                 DBT *key, DBT *data, u_int32_t flags)
+                 DBT *key, DBT *data, u_int32_t flags, int odhready)
 {
     DBT data2;
     void *mallocmem;
     int rc, updateid = -1, ipu = ip_updates_enabled(bdb_state);
     unsigned long long *genptr = NULL;
-
-#if 0
-   printf("bdb_put_pack\n");
-   fsnapf(stdout, key->data, key->size);
-   sleep(1);
-#endif
 
     if (ipu && key->size >= 8) {
         genptr = (unsigned long long *)key->data;
@@ -1208,13 +1197,11 @@ int bdb_put_pack(bdb_state_type *bdb_state, int is_blob, DB *db, DB_TXN *tid,
 
     rc = bdb_prepare_put_pack_updateid(
         bdb_state, is_blob, data, &data2, updateid, &mallocmem,
-        ALLOC_STACKBUF(data->size + ODH_SIZE_RESERVE));
+        ALLOC_STACKBUF(data->size + ODH_SIZE_RESERVE), odhready);
 
     if (rc == 0) {
         rc = db->put(db, tid, key, &data2, flags);
-        if (mallocmem) {
-            free(mallocmem);
-        }
+        free(mallocmem);
     }
 
     if (ipu && updateid >= 0 && key->size >= 8) {
@@ -1261,7 +1248,7 @@ int bdb_cput_pack(bdb_state_type *bdb_state, int is_blob, DBC *dbcp, DBT *key,
 
     rc = bdb_prepare_put_pack_updateid(
         bdb_state, is_blob, data, &data2, updateid, &mallocmem,
-        ALLOC_STACKBUF(data->size + ODH_SIZE_RESERVE));
+        ALLOC_STACKBUF(data->size + ODH_SIZE_RESERVE), 0);
 
     if (rc == 0) {
         rc = dbcp->c_put(dbcp, key, &data2, flags);
@@ -1376,4 +1363,62 @@ inline void bdb_cleanup_fld_hints(bdb_state_type *bdb_state)
         free(bdb_state->fld_hints);
         bdb_state->fld_hints = NULL;
     }
+}
+
+inline void bdb_set_logical_live_sc(bdb_state_type *bdb_state)
+{
+    if (bdb_state == NULL) {
+        logmsg(LOGMSG_ERROR, "%s(NULL)!!\n", __func__);
+        return;
+    }
+    bdb_state->logical_live_sc = 1;
+}
+
+inline void bdb_clear_logical_live_sc(bdb_state_type *bdb_state)
+{
+    if (bdb_state == NULL) {
+        logmsg(LOGMSG_ERROR, "%s(NULL)!!\n", __func__);
+        return;
+    }
+    bdb_state->logical_live_sc = 0;
+}
+
+int bdb_unpack_heap(bdb_state_type *bdb_state, void *in, size_t inlen,
+                    void **out, size_t *outlen, void **freeptr)
+{
+    struct odh odh;
+    int rc;
+    char odhd = bdb_state->ondisk_header;
+
+    /* Force ODH in case that a schema change
+       is removing the ODH from the table. */
+    bdb_state->ondisk_header = 1;
+    rc = bdb_unpack(bdb_state, in, inlen, NULL, 0, &odh, freeptr);
+    bdb_state->ondisk_header = odhd;
+
+    if (rc != 0) {
+        *out = NULL;
+        free(*freeptr);
+    } else {
+        *outlen = odh.length;
+        *out = odh.recptr;
+    }
+
+    return rc;
+}
+
+int bdb_pack_heap(bdb_state_type *bdb_state, void *in, size_t inlen, void **out,
+                  size_t *outlen, void **freeptr)
+{
+    uint32_t recsz;
+    int rc;
+
+    if (!bdb_state->ondisk_header)
+        return -1;
+    struct odh odh;
+    init_odh(bdb_state, &odh, in, inlen, 1);
+    rc = bdb_pack(bdb_state, &odh, NULL, 0, out, &recsz, freeptr);
+    if (rc == 0)
+        *outlen = recsz;
+    return rc;
 }