--- a/bdb/odh.c
+++ b/bdb/odh.c
@@ -1365,24 +1365,6 @@ inline void bdb_cleanup_fld_hints(bdb_state_type *bdb_state)
     }
 }
 
-inline void bdb_set_logical_live_sc(bdb_state_type *bdb_state)
-{
-    if (bdb_state == NULL) {
-        logmsg(LOGMSG_ERROR, "%s(NULL)!!\n", __func__);
-        return;
-    }
-    bdb_state->logical_live_sc = 1;
-}
-
-inline void bdb_clear_logical_live_sc(bdb_state_type *bdb_state)
-{
-    if (bdb_state == NULL) {
-        logmsg(LOGMSG_ERROR, "%s(NULL)!!\n", __func__);
-        return;
-    }
-    bdb_state->logical_live_sc = 0;
-}
-
 int bdb_unpack_heap(bdb_state_type *bdb_state, void *in, size_t inlen,
                     void **out, size_t *outlen, void **freeptr)
 {