--- a/db/config.c
+++ b/db/config.c
@@ -341,6 +341,7 @@ static char *legacy_options[] = {"disallow write from beta if prod",
                                  "reorder_socksql_no_deadlock off",
                                  "disable_tpsc_tblvers",
                                  "on disable_etc_services_lookup",
+                                 "off osql_odh_blob",
                                  "legacy_schema on"};
 int gbl_legacy_defaults = 0;
 int pre_read_legacy_defaults(void *_, void *__)