--- a/tests/basic.test/off_osql_odh_blob.testopts
+++ b/tests/basic.test/off_osql_odh_blob.testopts
@@ -1,0 +1,1 @@
+off osql_odh_blob