… bdb/upd.c
… db/glue.c
… db/localrep.c
… db/osqlblockproc.c
… db/osqlcomm.c
… db/osqlcomm.h
… db/osqlshadtbl.c
… db/osqlsqlthr.c
… db/record.c
… db/tag.c
… db/types.h
… schemachange/sc_records.c
… schemachange/schemachange.c
… tests/odh_blobs.test/Makefile
… tests/odh_blobs.test/expected
… tests/odh_blobs.test/runit
… tests/odh_blobs.test/t1.csc2
… tests/odh_blobs.test/t2.csc2
… tests/odh_blobs.test/t3.csc2
… tests/sc_add_vutf8.test/off_osql_odh_blob.testopts
… tests/sc_vutf8_blob2.test/off_osql_odh_blob.testopts
… tests/snapblob.test/off_osql_odh_blob.testopts
… tests/tunables.test/t00_all_tunables.expected
… tests/vutf8_dbstore.test/off_osql_odh_blob.testopts
… tests/vutf8_rowlocks.test/off_osql_odh_blob.testopts