Instead of sending a blob to the master as is, pack and compress the blob on a replicant, and send the preprocessed blob to the master.
This approach significantly reduces the network IO between replicants and the master, as well as the memory footprint and CPU use incurred by compression on the master.

The chart below demonstrates how it works.

```
+-----------+                          +-----------+
| Replicant |                          |   Master  |
+-----------+                          +-----------+
     |                                       |
odhfy_blob_buffer()                          |
     |                                       |
     +--- Add ODH                            |
     +--- Compress                           |
     +--- Mask OSQL_BLOB_ODH_BIT             |
     |    in blob_buffer_t.odhind            |
     |                                       |
osql_send_qblob() -------------------> osql_process_packet()
                                             |
                                             ~
                                             |
                                       bdb_pack(..., odhready=(odhind&OSQL_BLOB_ODH_BIT))
                                             |
                                       is odhready true?
                                            / \
                                        Y  /   \ N
                                          /     \
                                 Use as is       +--- Add ODH
                                                 +--- Compress
```

If a schema changing is modifying the compression algorithm of the blob field, or creating a new expressional index on the blob column, the proprocessed blob must be unpacked on the master.