@ldennis 2019-01-31
Looking at record.c, does this work for sp? And we may need more changes internally for oplog based localrep to work? By the way, Mark's physical rep should just work. :+1: