@riverszhang89 2019-01-31
It should work inside a SP - `dbtable:insert` internally forms an INSERT statement and passes the statement down to SQLite.

Blobs are "unodhfied" (remove ODH and decompress) first and then are saved into oplog (see [here](https://github.com/bloomberg/comdb2/pull/1399/files#diff-20b8ebf66c0ff339c2b20563d1f0d548R116)) . So it should also work with cdb2rcv.