@ldennis 2019-01-31
What about trigger? Do we unodhfied before adding to queue? I was looking for that logic.