@riverszhang89 2019-02-08
## Benchmarks

Write test: Inserting a blob into a LZ4-compressed blob column
Read test: Selecting a pre-processed blob from the shadow table

### Results

| Dataset | Original Size | Compressed Size | Compression Ratio | Write Test Runtime Percent of Change Comparing to Master | Read Test Runtime Percent of Change Comparing to Master
|----------|----------|------|---------|----------------------------------------------------------|--|
| [NYC Bus Delay Open Data](https://data.cityofnewyork.us/Transportation/Bus-Breakdown-and-Delays/ez4e-fazm)|66MiB|18MiB|26.63%| -32% | -77% |
| Comdb2 github repo | 102MiB | 57MiB | 55.95% | -22% | -52% |
| /dev/urandom | 67MiB | 67MiB | 100% | -3%|-5%|