@ldennis 2019-02-08
So reading compressed blobs out from shadow is even faster? whoah ..