@riverszhang89 2019-02-08
Yeah - Looks like the overhead of saving lengthy records is even greater than the overhead of compression. I will do another benchmark on small blobs.