@ldennis 2019-02-08
That makes sense actually.. Temp table needs to do disk io for that size as I think temp table cache is definitely smaller than that ..