@riverszhang89 2019-02-08
### Benchmarks Ⅱ

| Dataset | Original Size | Compressed Size | Compression Ratio | Write Test Runtime Change (%) | Read Test Runtime Change (%)
|----------|----------|------|---------|----------------------------------------------------------|--|
| [OATH Trials Division Case Status](https://data.cityofnewyork.us/City-Government/OATH-Trials-Division-Case-Status/y3hw-z6bm)|176KiB|56KiB|29.56%| -77% | +20% |
| /dev/urandom |180KiB | 180KiB | 100% | -8%|+29%|

Reading pre-processed small blobs from the shadow table is slower. However the runtime is neglectable - it is sub-millisecond.