@riverszhang89 2019-02-14
Writes are still much faster for small blobs after pre-processing. In addition, constantly reading from a shadow table isn't a common use case. In real life, shadow tables are merged with data from the real btrees. The overhead of reading pre-processed blobs is just a small portion.